[PLG-2080]: Register creditExpiryHandler with next gen
