+++ 120-ng-manager/src/main/java/io/harness/ng/NextGenApplication.java
@@ -65,7 +65,7 @@ import io.harness.connector.ConnectorDTO;
 import io.harness.connector.entities.Connector;
 import io.harness.connector.gitsync.ConnectorGitSyncHelper;
 import io.harness.controller.PrimaryVersionChangeScheduler;
-import io.harness.credit.schedular.CreditExpiryHandler;
+import io.harness.credit.schedular.CreditExpiryIteratorHandler;
 import io.harness.enforcement.client.CustomRestrictionRegisterConfiguration;
 import io.harness.enforcement.client.RestrictionUsageRegisterConfiguration;
 import io.harness.enforcement.client.custom.CustomRestrictionInterface;
@@ -645,7 +645,7 @@ public class NextGenApplication extends Application<NextGenConfiguration> {
     injector.getInstance(GitlabSCMOAuthTokenRefresher.class)
         .registerIterators(ngIteratorsConfig.getOauthTokenRefreshIteratorConfig().getThreadPoolSize());
     injector.getInstance(NGVaultUnsetRenewalHandler.class).registerIterators(5);
-    injector.getInstance(CreditExpiryHandler.class).registerIterator(2);
+    injector.getInstance(CreditExpiryIteratorHandler.class).registerIterator(2);
   }
 
   public void registerJobs(Injector injector) {
