+++ 930-ng-license-manager/src/main/java/io/harness/credit/entities/CICredit.java
@@ -10,6 +10,7 @@ import static io.harness.annotations.dev.HarnessTeam.GTM;
 
 import io.harness.annotations.StoreIn;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.iterator.PersistentRegularIterable;
 import io.harness.ng.DbAliases;
 
 import dev.morphia.annotations.Entity;
@@ -28,14 +29,22 @@ import org.springframework.data.annotation.TypeAlias;
 
 @Persistent
 @TypeAlias("io.harness.credit.entities.CICredit")
-public class CICredit extends Credit {
-    @Override
-    public Long obtainNextIteration(String fieldName) {
-        return null;
+public class CICredit extends Credit implements PersistentRegularIterable {
+  @Override
+  public Long obtainNextIteration(String fieldName) {
+    if (CreditsKeys.creditExpiryCheckIteration.equals(fieldName)) {
+      return creditExpiryCheckIteration;
     }
+    throw new IllegalArgumentException("Invalid fieldName " + fieldName);
+  }
 
-    @Override
-    public String getUuid() {
-        return null;
-    }
+  @Override
+  public String getUuid() {
+    return this.id;
+  }
+
+  @Override
+  public void updateNextIteration(String fieldName, long nextIteration) {
+    this.creditExpiryCheckIteration = nextIteration;
+  }
 }
